Implement game inventory system
This system is built around a list where items can be added as the
player progresses through the game. These games are stored in a bag
which you can check to see what your games are, and optionally play
them.
It allows for more games to be implemented.
No game disc is shown for now, but will do it soon.